move productAnalytics repo to DAL
