+++ backend/src/services/productAnalyticsService.ts
@@ -1,18 +1,22 @@
 import { LoggerBase } from '@crowd/logging'
+import { PgPromiseQueryExecutor } from '@crowd/data-access-layer/src/queryExecutor'
+import { createSession, updateSession, createEvent } from '@crowd/data-access-layer/src/productAnalytics'
 import { IServiceOptions } from './IServiceOptions'
-import ProductAnalyticsRepository from '@/database/repositories/productAnalyticsRepository'
 
 export default class ProductAnalyticsService extends LoggerBase {
+  private readonly qx: PgPromiseQueryExecutor
+
   options: IServiceOptions
 
   constructor(options: IServiceOptions) {
     super(options.log)
     this.options = options
+    this.qx = new PgPromiseQueryExecutor(this.options.productDb)
   }
 
   public async createSession(data) {
     try {
-      return await ProductAnalyticsRepository.createSession(data, this.options)
+      return await createSession(this.qx, data)
     } catch (error) {
       throw new Error('Error during session create!')
     }
@@ -20,7 +24,7 @@ export default class ProductAnalyticsService extends LoggerBase {
 
   public async updateSession(id, data) {
     try {
-      return await ProductAnalyticsRepository.updateSession(id, data, this.options)
+      return await updateSession(this.qx, id, data)
     } catch (error) {
       throw new Error('Error during session update!')
     }
@@ -28,7 +32,7 @@ export default class ProductAnalyticsService extends LoggerBase {
 
   public async createEvent(data) {
     try {
-      return await ProductAnalyticsRepository.createEvent(data, this.options)
+      return await createEvent(this.qx, data)
     } catch (error) {
       throw new Error('Error during event create!')
     }
